rename mir -> thir around abstract consts
